requirements.txt
python-telegram-bot~=21.6
setuptools~=69.5.1